chore: clean up codebase and remove temporary files
- Remove test files from root directory
- Remove backup files (.bak, .bak2, .bak3)
- Remove build artifacts and output files
- Remove disabled source files (.disabled)
- Remove temporary CTest directories
- Clean up neo-data directory
- Add CI/CD configurations and documentation

+++ src/network/p2p/payloads/addr_payload.cpp
@@ -17,11 +17,11 @@ void AddrPayload::SetAddressList(const std::vector<NetworkAddressWithTime>& addr
     addressList_ = addresses;
 }
 
-int AddrPayload::GetSize() const
+size_t AddrPayload::GetSize() const
 {
     // Calculate the size based on addresses
     // This matches the C# implementation: AddressList.GetVarSize()
-    int size = 0;
+    size_t size = 0;
 
     // VarInt for count
     if (addressList_.size() < 0xFD)
